+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - refactor: error log to throw error
 - refactor: util method
 - refactor: move API `process` to `workflow`
-- feat: `getHelper` API
+- feat: helper `get` API, remove `language` API
+- feat: config `set` and `get` API
 
 ## 3.2.1